==> 09_12_signature_TASA_predictive_modeling_SHIP_0.85_prevalence.Rmd <==
---
title: "MetaPac"
subtitle: "12 signature TASA (4 met arm B predictive performance)"
author: "_umahajan_"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_notebook:
    theme: united
    highlight: tango
    css: custom.css
    number_sections: true
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
      smooth_scroll: true
    code_folding: show
---

```{r setup, include=FALSE}
chooseCRANmirror(graphics = TRUE, ind = 1)
knitr::opts_chunk$set(
  tidy.opts = list(width.cutoff = 85),
  tidy = TRUE,
  echo = TRUE,
  warning = FALSE,
  message = FALSE,
  fig.align = "center",
  tidy.opts = list(
    indent = 2,          
    width.cutoff = 60),
  comment = "#>"
)
options(width = 60,
        digits = 3,
        scipen = 999)
```

# Load packages and datasets
```{r packages}
rm(list = ls())
##---------------------------------------------------------------
##                      required packages                      --
##---------------------------------------------------------------
scriptLibraries <-  c("here",
                      "tidyverse",
                      "h2o",
                      "caret",
                      "glmnet",
                      "Rmisc",
                      "ROCR")
##---------------------------------------------------------------
##                      load functions                    --
##---------------------------------------------------------------
source("https://raw.githubusercontent.com/umahajanatlmu/useful_commands/main/auxillary/basicFunctions.R")
##---------------------------------------------------------------
##                        load packages                        --
##---------------------------------------------------------------
installScriptLibs(scriptLibraries)
##---------------------------------------------------------------
##                    set working directory                    --
##---------------------------------------------------------------
here::here()
```

## initiate h2o environment
```{r}
##----------------------------------------------------------------
##             detect the number of cores available             --
##----------------------------------------------------------------
myCores = parallel::detectCores(all.tests = TRUE) - 1

if (myCores > 20) {
  myCores = 20
} else
  myCores = myCores


memFreeG = 50
#Sys.setenv(JAVA_HOME = "/Library/Java/JavaVirtualMachines/zulu-11.jdk")
##----------------------------------------------------------------
##                         initiate h2o                         --
##----------------------------------------------------------------
h2o.init(
  nthreads = myCores,
  min_mem_size = paste(memFreeG, "g", sep = ""),
  max_mem_size = paste(memFreeG, "g", sep = "")
)
```
# load SHIP metadata
```{r}
set.seed(1234456) 

ship_metadata <- readxl::read_excel("../../results/samples_selection_SHIP.xlsx")

ship_metadata <- ship_metadata %>%
  dplyr::select(NODM_SHIP_TREND1, DM_PDAC_SHIP_TREND1) %>%
  drop_na(NODM_SHIP_TREND1) %>%
  mutate(PT = as.character(NODM_SHIP_TREND1),
         DM_PDAC_SHIP_TREND1 = as.character(DM_PDAC_SHIP_TREND1))

pdac_cases <- na.omit(ship_metadata$DM_PDAC_SHIP_TREND1)

ship_metadata$Disease_status <- ifelse(ship_metadata$NODM_SHIP_TREND1 %in% pdac_cases, "PDAC", "CP")

mrt_t1_pank <- haven::read_sav("../../SHIP_2020_160_M+D/data/ship_2020_160_md_t1.sav")
mrt_t1_pank$zz_nr <- as.character(mrt_t1_pank$zz_nr)

mrt_t1_pank <- mrt_t1_pank[mrt_t1_pank$zz_nr %in% ship_metadata$PT, ]

mrt_t1_pank <- mrt_t1_pank %>%
  dplyr::select(sex_trend1, age_trend1,ipw) %>%
  mutate(age_trend1 = as.numeric(age_trend1),
         ipw = as.numeric(ipw))

summary(mrt_t1_pank)

table(mrt_t1_pank$sex_trend1)
## simulate values in a range
# Number of samples for each category
n_cp <- table(ship_metadata$Disease_status)[[1]]
n_pdac <- table(ship_metadata$Disease_status)[[2]]

# Simulate CA19.9 levels for CP and PDAC

for (r in 1:nrow(ship_metadata)) {
  if (ship_metadata[r, "Disease_status"] == "PDAC") {
    ship_metadata[r, "CA19_9"] <-  mean(runif(10, min = 39.1, max = 1036.0))
  } else {
    ship_metadata[r, "CA19_9"] <-  mean(runif(10, min = 4.2, max = 28.0))
  }
    
}


ship_metadata <- ship_metadata %>%
  dplyr::select(PT, Disease_status, CA19_9)

### process PT data
lookup_table <- readxl::read_excel("../../SHIP_2020_160_M+D/Sample_matching_SHIP_ÜbergabeOehrle20032024.xlsx")
ship_metadata$PT <- as.character(lookup_table$Auftragsnummer[match(ship_metadata$PT, lookup_table$ZZ_NR)])
```

# get metabolome data from SHIP
```{r}
prevalence <- 0.85/100

clp_dat <- readxl::read_xlsx("../../../../MetaPac_SHIP_results/KLUM-0102-22CLP/KLUM-0102-22CLP COMPLEX LIPID DATA TABLE.XLSX",
                             sheet = "Species Concentrations",
                             skip = 17)

clp_dat_names <- readxl::read_xlsx("../../../../MetaPac_SHIP_results/KLUM-0102-22CLP/KLUM-0102-22CLP COMPLEX LIPID DATA TABLE.XLSX",
                                   sheet = "Species Concentrations",
                                   skip = 9)
colnames(clp_dat) <- colnames(clp_dat_names)

clp_dat <- clp_dat %>%
  dplyr::select(-`CLIENT SAMPLE ID`, -...1) %>%
  column_to_rownames("...2") %>%
  mutate_all(as.numeric) %>%
  sjmisc::rotate_df() 

## subset clp data for 4 met signature
n12_met_id <- readxl::read_excel("../../../../12_met_signatures_id/metabolites_to_ids.xlsx")

n4_met_id <- unique(na.omit(n12_met_id$Metabolon_CLP_ID))

clp_dat <- clp_dat %>%
  dplyr::select(any_of(n4_met_id))

colnames(clp_dat) <- n12_met_id$Metabolite_ID[match(colnames(clp_dat), n12_met_id$Metabolon_CLP_ID)]

banner("performance with simulated CA19.9")

processed_data <- clp_dat %>%
  rownames_to_column("PT") %>%
  inner_join(ship_metadata, by = "PT")

processed_data <- metapacR::ImputeTransformScale(processed_data,
                                                 Impute = TRUE,
                                                 Transform = TRUE,
                                                 Scaling = TRUE,
                                                 ScaleType = "Auto")

banner("4 met signature")
val_h2o <- as.h2o(processed_data)

## H2O machine learning model
model_4mets <- h2o.upload_mojo("../../ID_VD1_VD2_models/production_models/GLM_1_AutoML_20211203_140114.zip")

perf <- h2o.performance(model_4mets, val_h2o)
#print(perf)

cutoff_4mets <- 0.773555
# cutoff.selected.model <- 0.773555
## confusion matrix
cm <- as.data.frame(h2o.confusionMatrix(perf,cutoff_4mets))
## enlist categories
lvs <- c("CP", "PDAC")
## truth
truth <- factor(rep(lvs, times = c(cm$CP[1] + cm$PDAC[1], cm$CP[2] + cm$PDAC[2])), levels = rev(lvs))
## pred
pred <- factor(c(rep(lvs, times = c(cm$CP[1], cm$PDAC[1])), rep(lvs, times = c(cm$CP[2], 
                                                                               cm$PDAC[2]))), levels = rev(lvs))
## xtab
xtab <- table(pred, truth)
## confusion matrix
cm <- caret::confusionMatrix(xtab, positive = "PDAC")
print(cm)

## add predictions to merged_data
prediction <- as.data.frame(h2o.predict(model_4mets, val_h2o))
colname <- "prediction_with_simulation"

processed_data[[colname]] <- prediction$PDAC
#X_sub$prediction_4mets <- prediction$PDAC

cv.results <- data.frame()
# split data for 10 fold cross validation ------------ Create 10 equally size folds
folds <- cut(seq(1, nrow(val_h2o)), breaks = 10, labels = FALSE)
# performance on 10 fold cross validation ----------------
for (i in 1:10) {
  # Segement your data by fold using the which() function
  Indexes <- which(folds == i, arr.ind = TRUE)
  test.cv <- val_h2o[-Indexes, ]
  perf.cv <- h2o.performance(model_4mets, newdata = test.cv)
  
  cm <- as.data.frame(h2o.confusionMatrix(perf.cv, cutoff_4mets))
  ## enlist categories
  lvs <- c("CP", "PDAC")
  ## truth
  truth <- factor(rep(lvs, times = c(cm$CP[1] + cm$PDAC[1], cm$CP[2] + cm$PDAC[2])), levels = rev(lvs))
  ## pred
  pred <- factor(c(rep(lvs, times = c(cm$CP[1], cm$PDAC[1])), rep(lvs, times = c(cm$CP[2],
                                                                                 cm$PDAC[2]))), levels = rev(lvs))
  ## xtab
  xtab <- table(pred, truth)
  ## confusion matrix
  cm <- caret::confusionMatrix(xtab, positive = "PDAC", prevalence = prevalence)
  
  cv.results[i, 1] <- i
  cv.results[i, 2] <- perf.cv@metrics$MSE
  cv.results[i, 3] <- perf.cv@metrics$RMSE
  cv.results[i, 4] <- perf.cv@metrics$r2
  cv.results[i, 5] <- perf.cv@metrics$logloss
  cv.results[i, 6] <- perf.cv@metrics$AUC
  cv.results[i, 7] <- perf.cv@metrics$pr_auc
  cv.results[i, 8] <- perf.cv@metrics$Gini
  cv.results[i, 9] <- perf.cv@metrics$mean_per_class_error
  cv.results[i, 10] = cm$overall["Accuracy"]
  cv.results[i, 11] = cm$byClass["Specificity"]
  cv.results[i, 12] = cm$byClass["Sensitivity"]
  cv.results[i, 13] = cm$byClass["Pos Pred Value"]
  cv.results[i, 14] = cm$byClass["Neg Pred Value"]
}

colnames(cv.results) <- c("fold", "MSE", "RMSE", "R2", "logloss", "AUC", "PRAUC", "Gini", 
                          "Mean_per_class_error", "accuracy", "specificity", "sensitivity", "ppv", "npv")

cv.results.summary <- 
  cv.results[, !colnames(cv.results) %in% "fold"] %>% gather(factor_key = TRUE) %>% 
  group_by(key)%>% mutate(value = as.numeric(value)) %>%
  dplyr::summarise(mean = mean(value, na.rm = TRUE), 
                   sd = sd(value, na.rm = TRUE), 
                   max = max(value, na.rm = TRUE), 
                   min = min(value, na.rm = TRUE), 
                   n  = n(),
                   se = sd / sqrt(n),
                   lower.ci = mean - qt(1 - (0.05 / 2), n - 1) * se,
                   upper.ci = mean + qt(1 - (0.05 / 2), n - 1) * se)
print(cv.results.summary)


banner("performance without CA19.9")

banner("4 met signature")
val_h2o <- as.h2o(processed_data[, colnames(processed_data) %in% c(colnames(clp_dat), "Disease_status")])

## H2O machine learning model
model_4mets <- h2o.upload_mojo("../../ID_VD1_VD2_models/production_models/GLM_1_AutoML_20211203_140114.zip")

perf <- h2o.performance(model_4mets, val_h2o)
#print(perf)

cutoff_4mets <- 0.773555
# cutoff.selected.model <- 0.773555
## confusion matrix
cm <- as.data.frame(h2o.confusionMatrix(perf,cutoff_4mets))
## enlist categories
lvs <- c("CP", "PDAC")
## truth
truth <- factor(rep(lvs, times = c(cm$CP[1] + cm$PDAC[1], cm$CP[2] + cm$PDAC[2])), levels = rev(lvs))
## pred
pred <- factor(c(rep(lvs, times = c(cm$CP[1], cm$PDAC[1])), rep(lvs, times = c(cm$CP[2], 
                                                                               cm$PDAC[2]))), levels = rev(lvs))
## xtab
xtab <- table(pred, truth)
## confusion matrix
cm <- caret::confusionMatrix(xtab, positive = "PDAC")
print(cm)

## add predictions to merged_data
prediction <- as.data.frame(h2o.predict(model_4mets, val_h2o))
colname <- "prediction_without_simulation"

processed_data[[colname]] <- prediction$PDAC
#X_sub$prediction_4mets <- prediction$PDAC

cv.results <- data.frame()
# split data for 10 fold cross validation ------------ Create 10 equally size folds
folds <- cut(seq(1, nrow(val_h2o)), breaks = 10, labels = FALSE)
# performance on 10 fold cross validation ----------------
for (i in 1:10) {
  # Segement your data by fold using the which() function
  Indexes <- which(folds == i, arr.ind = TRUE)
  test.cv <- val_h2o[-Indexes, ]
  perf.cv <- h2o.performance(model_4mets, newdata = test.cv)
  
  cm <- as.data.frame(h2o.confusionMatrix(perf.cv, cutoff_4mets))
  ## enlist categories
  lvs <- c("CP", "PDAC")
  ## truth
  truth <- factor(rep(lvs, times = c(cm$CP[1] + cm$PDAC[1], cm$CP[2] + cm$PDAC[2])), levels = rev(lvs))
  ## pred
  pred <- factor(c(rep(lvs, times = c(cm$CP[1], cm$PDAC[1])), rep(lvs, times = c(cm$CP[2],
                                                                                 cm$PDAC[2]))), levels = rev(lvs))
  ## xtab
  xtab <- table(pred, truth)
  ## confusion matrix
  cm <- caret::confusionMatrix(xtab, positive = "PDAC", prevalence = prevalence)
  
  cv.results[i, 1] <- i
  cv.results[i, 2] <- perf.cv@metrics$MSE
  cv.results[i, 3] <- perf.cv@metrics$RMSE
  cv.results[i, 4] <- perf.cv@metrics$r2
  cv.results[i, 5] <- perf.cv@metrics$logloss
  cv.results[i, 6] <- perf.cv@metrics$AUC
  cv.results[i, 7] <- perf.cv@metrics$pr_auc
  cv.results[i, 8] <- perf.cv@metrics$Gini
  cv.results[i, 9] <- perf.cv@metrics$mean_per_class_error
  cv.results[i, 10] = cm$overall["Accuracy"]
  cv.results[i, 11] = cm$byClass["Specificity"]
  cv.results[i, 12] = cm$byClass["Sensitivity"]
  cv.results[i, 13] = cm$byClass["Pos Pred Value"]
  cv.results[i, 14] = cm$byClass["Neg Pred Value"]
}

colnames(cv.results) <- c("fold", "MSE", "RMSE", "R2", "logloss", "AUC", "PRAUC", "Gini", 
                          "Mean_per_class_error", "accuracy", "specificity", "sensitivity", "ppv", "npv")

cv.results.summary <- 
  cv.results[, !colnames(cv.results) %in% "fold"] %>% gather(factor_key = TRUE) %>% 
  group_by(key)%>% mutate(value = as.numeric(value)) %>%
  dplyr::summarise(mean = mean(value, na.rm = TRUE), 
                   sd = sd(value, na.rm = TRUE), 
                   max = max(value, na.rm = TRUE), 
                   min = min(value, na.rm = TRUE), 
                   n  = n(),
                   se = sd / sqrt(n),
                   lower.ci = mean - qt(1 - (0.05 / 2), n - 1) * se,
                   upper.ci = mean + qt(1 - (0.05 / 2), n - 1) * se)
print(cv.results.summary)
```

## boxplot

```{r}
## BMI plot
bmi_dat <- haven::read_sav("../../../../MetaPac-SHIP/SHIP_2024_39_D/SHIP_2024_39_D_T1_20240703.sav")
bmi_dat$zz_nr <- as.character(bmi_dat$zz_nr)

bmi_dat$PT <- as.character(lookup_table$Auftragsnummer[match(bmi_dat$zz_nr, lookup_table$ZZ_NR)])

processed_data <- bmi_dat %>%
  drop_na(PT) %>%
  right_join(processed_data, bmi_dat, by="PT") %>%
  rename(c("bmi_t1" = "BMI")) 

df <- processed_data %>%
  mutate(BMI_cat = case_when(
    BMI < 18.5 ~ "Underweight",
    BMI >= 18.5 & BMI <= 24.9 ~ "Normal",
    BMI >= 25.0 & BMI <= 29.9 ~ "Overweight",
    BMI >= 30.0 ~ "Obese"
  )) %>%
  dplyr::select(Disease_status, prediction_with_simulation, prediction_without_simulation, BMI, BMI_cat) %>%
  pivot_longer(!c(Disease_status, BMI, BMI_cat), names_to = "Category", values_to = "prediction") %>%
  mutate(BMI = as.numeric(BMI))

df$BMI_cat <- factor(df$BMI_cat, levels = c("Underweight","Normal","Overweight","Obese"))

## plot boxplot for 4 met signature
p <- ggplot(df,
            aes(x= Disease_status,
                y= prediction, 
                fill = Disease_status)) +
  geom_boxplot(color="black", show.legend = FALSE, alpha = 0.8) +
   scale_fill_manual(values = c("#7876B1FF","#BC3C29FF")) +
  ggnewscale::new_scale_fill() + 
  ggbeeswarm::geom_quasirandom(aes(fill= BMI_cat),
                               shape=21,
                               color="black", 
                               alpha = 0.8, 
                               method = "pseudorandom",
                               size = 1) +
  scale_fill_manual(values = rev(RColorBrewer::brewer.pal(4, "RdBu"))) +
  theme_bw() +
  geom_hline(yintercept = cutoff_4mets, size = 1, color = "black") +
  ggtitle("Performance of m-Metabolome Signature") +
  ggpubr::stat_compare_means(data=df, 
                     aes(x= Disease_status, 
                         y= prediction), 
                     ref.group = "CP",
                     method="t.test",
                     label = "..p..",
                     show.legend = FALSE) +
  xlab("") +
  ylab("m-Metabolome Signature") +
  facet_wrap(~Category)


  # scale_fill_gradientn(
  #         colours = rev(RColorBrewer::brewer.pal(10, "RdBu")),
  #         oob = scales::squish,
  #         limits = c(18.5, 35)
  #       ) +
  #       guides(fill = guide_colourbar(
  #         barwidth = unit(0.3, "cm"),
  #         ticks.colour = "black",
  #         frame.colour = "black"
  #       ))

print(p)

ggsave(filename = "results/boxPlot_4met_signature_ship.pdf",
       plot=p,
       width = 5, height =4,
       dpi= 300)


p <- ggplot(df,
            aes(x= BMI,
                y= prediction, 
                fill = Disease_status)) +
  geom_point(shape=21)+
  geom_smooth() +
  facet_wrap(~Category)
  

h2o.shutdown(prompt = F)
```
## demographics
```{r}
mrt_t1_pank <- haven::read_sav("../../SHIP_2020_160_M+D/data/ship_2020_160_md_t1.sav")
mrt_t1_pank$zz_nr <- as.character(mrt_t1_pank$zz_nr)

mrt_t1_pank <- mrt_t1_pank[mrt_t1_pank$zz_nr %in% ship_metadata$PT, ]
```

# Computing environment
```{r}
sessionInfo()
```
